Update the placeholder after the textView has loaded
When setting a default placeholder and an attributedString before `_textKitComponents.textView` was created, but the placeholder and the string were appearing in the text node.

`_updateDisplayingPlaceholder` is called in `setAttributedString`, but since  `_textKitComponents.textView` was nil the placeholder was still told to displaying. Calling `_updateDisplayingPlaceholder` after `_textKitComponents.textView` loads fixes this.