Add @synchronized around UIImage draw to prevent concurrently decoding the same instance.
This is a workaround necessary due to an iOS 9 bug.  I'm not gating it to iOS 9 only so as to
not exaggerate iOS version-specific behavior differences in the framework, since a majority are
on iOS 9 anyway.  If we can confirm the bug is fixed in a later iOS version, then it will be gated.

Issue tracked in greater detail here: https://github.com/facebook/AsyncDisplayKit/issues/1068