Fix custom class drawing methods are not called for ASTextNode and ASImageNode
ASTextNode and ASImageNode currently implement drawing methods as instance methods. This breaks the custom class drawing methods. They will never be called for a subclass. Fixing it by adding a delegate object to the text / image node drawing methods to be able to dispatch the call to the instance method in ASDisplayNode.

To be able to work this way the original parameters needs to be passed to the superclass if a subclass calls the superclasses class drawing method.