Refactor ASTextNode
The main goal of the refactoring was to reduce state that we are currently hold in the ASTextNode and all the Core Text cmponents we use . In the ASTextNode we are holding on the renderer object what can result in weird behavior and we also need to be careful when to invalidate the renderer or pass new state down to the Core Text components.

The basic idea is that neither the ASTextNode nor layout should ever have a strong reference to the renderer object. This is to reduce memory load when loading thousands and thousands of text nodes into memory at once. Instead
 we maintain a LRU renderer cache that is queried via stack-allocated keys.

The renderer will be shared across all ASTextNode's as well as reused if possible. Currently the renderer cache has a an arbitrary size of 500 renderers and cleans up 100 (20% / 0.2) if the cost goes over that limit.