[ASCollectionNode/ASTableNode] Remeasure cells on trait collection change
It isn’t enough to call `setNeedsLayout` on a cell since the container size will change on a trait collection change. We must call `measure` with the new container size as the constrainted size.

This is all happening on main, which I don’t like, but since we are reacting to a trait collection change does it make sense for it to happen anywhere else?