Revert "Set alignItems to ASStackLayoutAlignItemsStretch in ASStackLayoutSpec (#2187)"
This reverts commit 37416f16965a686264937366f78fd2f92b419c70.

We are reverting this commit only for Pinterest.  The new default of stretch will be maintained on
ASDK master, and when Pinterest syncs to this in 1-2 weeks, we'll communicate to the team about
the new default.

The small number of changes that were made to address code that was relying on Start (previous default)
will continue to work safely, because they'll just be re-setting the value to Start which is now default again.

After the full adoption of ASDK 2.0 APIs, the default will become Stretch again (for alignItems).