[Build] Restore operational AsyncDisplayKit-Prefix.pch
This was regressed by the large project file changes in 404795dc02aa57b167223bfaee7c380907022ca1

I don't yet have full context on the goals of that change, but it's definitely necessary to have the .pch
with the current requirements of the project. This line ensures that Cocoapods adds it.