[ASCollectionView] Allow mixing supplementaryViews and Nodes by relaxing assertions.
These changed recently with the addition of the ability to provide nodeBlocks.

It might be more strict to assert that nil can't be provided unless the datasource supports
inter-op, but I think it is also pretty reasonable to just use an empty ASCellNode in this case.

Either way, I'm really amazed / surprised to see all the improvements that appeared in ASDataController!